+++ src/modules/manufacturing/entities/apm-sp-requisition-line-mach.ts
@@ -32,6 +32,10 @@ export class SparePartReqLineMach extends BaseEntity {
   @Column({ name: 'QUANTITY' })
   quantity!: number;
 
+  @Field({ defaultValue: false })
+  @Column({ name: 'NON_KS' })
+  nonKS!: boolean;
+
   @Field()
   @Column({ name: 'CREATED_AT' })
   createdAt!: Date;