+++ src/modules/procurement/courier-fee-prepaid.in.ts
@@ -74,4 +74,11 @@ export class CourierFeePrepaidInput implements Partial<CourierFeePrepaid> {
 
   @Field({ nullable: true })
   emailUser?: string;
+
+  @Field({ nullable: true })
+  periodeBayar?: string;
+
+  @Field({ nullable: true })
+  @IsDate()
+  financeReceiptDate?: Date;
 }