+++ src/modules/procurement/entities/courier-fee-prepaid.ts
@@ -91,4 +91,12 @@ export class CourierFeePrepaid extends BaseEntity {
   @Field({ nullable: true })
   @Column({ name: 'EMAIL_USER' })
   emailUser?: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'PERIODE_BAYAR' })
+  periodeBayar?: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'FINANCE_RECEIPT_DATE' })
+  financeReceiptDate?: Date;
 }