+++ src/modules/warehouse/entities/atj-ims-inv-part-in-stock.ts
@@ -1,0 +1,74 @@
+import { Field, ObjectType } from 'type-graphql';
+import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
+
+@Entity('ATJ_IMS_INV_PART_IN_STOCK')
+@ObjectType()
+export class ImsInvPartInStock extends BaseEntity {
+  @Field()
+  @PrimaryColumn({ name: 'PART_NO' })
+  partNo!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'CONTRACT' })
+  contract!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'LOCATION_NO' })
+  locationNo!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'LOT_BATCH_NO' })
+  lotBatchNo!: string;
+
+  @Field()
+  @Column({ name: 'WAREHOUSE_ID' })
+  warehouseId!: string;
+
+  @Field()
+  @Column({ name: 'BAY_ID' })
+  bayId!: string;
+
+  @Field()
+  @Column({ name: 'ROW_ID' })
+  rowId!: string;
+
+  @Field()
+  @Column({ name: 'TIER_ID' })
+  tierId!: string;
+
+  @Field()
+  @Column({ name: 'BIN_ID' })
+  binId!: string;
+
+  @Field()
+  @Column({ name: 'QTY_ONHAND' })
+  qtyOnhand!: number;
+
+  @Field({ nullable: true })
+  @Column({ name: 'CATCH_QTY_ONHAND' })
+  catchQtyOnhand?: number;
+
+  @Field({ nullable: true })
+  @Column({ name: 'NOTE' })
+  note?: string;
+
+  @Field()
+  @Column({ name: 'IS_DELETED' })
+  isDeleted!: number;
+
+  @Field()
+  @Column({ name: 'CREATED_DATE' })
+  createdDate!: Date;
+
+  @Field()
+  @Column({ name: 'CREATED_BY' })
+  createdBy!: string;
+
+  @Field()
+  @Column({ name: 'MODIFIED_DATE' })
+  modifiedDate!: Date;
+
+  @Field()
+  @Column({ name: 'MODIFIED_BY' })
+  modifiedBy!: string;
+}